Simplify web setup related to libolm
During a setup of Twake Chat for web, I ran into the section about libolm.

It seems to be working when downloading olm.js from here https://gitlab.matrix.org/matrix-org/olm/-/packages/2190 instead of building it with Docker.

Why do you think ? It seems a lot simpler for me to setup, but I lack the necessary background and maybe it will not work everywhere/everytime ?